Add feed title, story scroll and load progress bar
Add a title bar layer 'class' used to show the feed on the headline menulist page.
Fix the layout of the story page on round watchfaces (may break formatting for
rectangular faces). Allow the story to flow and scroll correctly.
Provide a progress load bar for transfering the story from phone to watch. To
achieve this the buffer size is decreased to 128 bytes, which may need
tweaking.