Reduce inbox and outbox size
The buffer sizes cause problems for aplite watches due to lack of memory.
The code set them to app_message_inbox_size_maximum and
app_message_outbox_size_maximum which was 8k for SDK 4.0. This size also
seemed unecessary, so this reduces both to 440B. This should be enough
for the largest headline (160B) and the largest feed name (30B) as set in
main.js.
Also fix the wscript which was copying all of the javascript files into
the .pbw file, whereas it should only be pebble-js-app.js